Remove incomplete query/node/* metrics
While testing in production the exporter a bug came up,
namely that the query/node/* metrics are missing a label,
the 'server' one. These metrics are broker specific and are
meant to give more information to pin point latencies between
a broker and a historical to fulfill a query. This information
is redundand since the exporter already collects separated
latency/size metrics for broker and historical, and can always
be found in the logs emitted by default if a specific investigation
needs to be done.

Bug: T177459
Change-Id: Ic22cb3bda825c4b6d00a65aafc188d7bf19a2b15