Add support to pull datapoints from Kafka
This change introduces a separate thread that is able to pull
data from a Kafka topic and insert datapoints in the shared
queue that separate ingestion from processing.

The idea is to set Druid daemons that need to push hundreds
of datapoints/s to Kafka (via the KafkaEmitter) and collect them
at a slower pace in the Exporter.

GH issue: #11

Change-Id: Ibc82be5883f20c26b50342d2032381086bcd218a