Fix tests for multi-threading code
After the introduction of threads, tests have been failing/hanging.
The main issue seems to be a race condition in the tests due to checking
too soon if a datapoints has been saved correctly into internal data
structures. The bug seems to be solved simply adding a small delay in
the tests after registering a datapoint.

This change also adds a way to stop threads programmatically, to avoid
tests to hang when done.

GH #13

Change-Id: I707ab0afca5f149d053f1f92834e0a64a7d1b960